refactor: format code & move icon to components
