--- src/assets/CopyIcon.vue
@@ -1,27 +1,0 @@
-<template>
-	<svg
-		xmlns="http://www.w3.org/2000/svg"
-		viewBox="0 0 24 24"
-		:width="size"
-		:height="size"
-		stroke="currentColor"
-		stroke-width="1.5"
-		stroke-linecap="round"
-		stroke-linejoin="round"
-		fill="none"
-		shape-rendering="geometricPrecision"
-	>
-		<path
-			d="M8 17.929H6c-1.105 0-2-.912-2-2.036V5.036C4 3.91 4.895 3 6 3h8c1.105 0 2 .911 2 2.036v1.866m-6 .17h8c1.105 0 2 .91 2 2.035v10.857C20 21.09 19.105 22 18 22h-8c-1.105 0-2-.911-2-2.036V9.107c0-1.124.895-2.036 2-2.036z"
-		></path>
-	</svg>
-</template>
-
-<script lang="ts" setup>
-defineProps({
-  size: {
-		type: Number,
-		default: 14
-  }
-})
-</script>
