+++ src/composables/useCopyCode.ts
@@ -2,20 +2,20 @@ import { useClipboard } from '@vueuse/core'
 import { Ref } from 'vue'
 
 type CopyCodeParams = {
-	code: string
-	checkIconRef: Ref<boolean>
+  code: string
+  checkIconRef: Ref<boolean>
 }
 
 export const useCopyCode = async ({ code, checkIconRef }: CopyCodeParams) => {
-	const { copy } = useClipboard({
-		source: code
-	})
-	
-	checkIconRef.value = false
-	await copy(code)
-	checkIconRef.value = true
-	
-	setTimeout(() => {
-		checkIconRef.value = false
-	}, 1500)
+  const { copy } = useClipboard({
+    source: code
+  })
+
+  checkIconRef.value = false
+  await copy(code)
+  checkIconRef.value = true
+
+  setTimeout(() => {
+    checkIconRef.value = false
+  }, 1500)
 }
